cmds: minor: Fix types in 'record' command

+++ src/cmds/record/record.c
@@ -102,7 +102,7 @@ static int sin_callback(const void *inputBuffer, void *outputBuffer,
 
 	for (i = 0; i < framesPerBuffer; i++) {
 		double x = 2 * 3.14 * (i % _sin_w) / _sin_w;
-		int tmp = (1. + _sin(x)) * _sin_h;
+		uint16_t tmp = (uint16_t) ((1. + _sin(x)) * _sin_h);
 
 		in_buf[cur_ptr++] = tmp;
 		if (chan_n == 2) {
